Fix tests logging; do not initialize the search client if the search … (#12681)
* Fix tests logging; do not initialize the search client if the search config is not present

* Fix logging and parallelization defaults 1C

* Fix logging and parallelization defaults 1C

* Fix logging and parallelization defaults 1C

* Fix logging and parallelization defaults 1C

* Fix logging and parallelization defaults 1C

* Fix logging and parallelization defaults 1C

+++ openmetadata-service/src/main/java/org/openmetadata/service/elasticsearch/ElasticSearchIndexDefinition.java
@@ -72,20 +72,26 @@ public class ElasticSearchIndexDefinition {
   }
 
   public void createIndexes(ElasticSearchConfiguration esConfig) {
-    for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
-      searchClient.createIndex(elasticSearchIndexType, esConfig.getSearchIndexMappingLanguage().value());
+    if (searchClient != null) {
+      for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
+        searchClient.createIndex(elasticSearchIndexType, esConfig.getSearchIndexMappingLanguage().value());
+      }
     }
   }
 
   public void updateIndexes(ElasticSearchConfiguration esConfig) {
-    for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
-      searchClient.updateIndex(elasticSearchIndexType, esConfig.getSearchIndexMappingLanguage().value());
+    if (searchClient != null) {
+      for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
+        searchClient.updateIndex(elasticSearchIndexType, esConfig.getSearchIndexMappingLanguage().value());
+      }
     }
   }
 
   public void dropIndexes() {
-    for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
-      searchClient.deleteIndex(elasticSearchIndexType);
+    if (searchClient != null) {
+      for (ElasticSearchIndexType elasticSearchIndexType : ElasticSearchIndexType.values()) {
+        searchClient.deleteIndex(elasticSearchIndexType);
+      }
     }
   }
